reconstruct directory
