+++ bgmi/download.py
@@ -6,14 +6,15 @@ import os
 
 from bgmi.config import SAVE_PATH, DOWNLOAD_DELEGATE
 from bgmi.models import STATUS_DOWNLOADING, STATUS_NOT_DOWNLOAD, Download
-from bgmi.services import XunleiLixianDownload, Aria2DownloadRPC, RRDownload, TransmissionRPC
+from bgmi.downloader.aria2_rpc import Aria2DownloadRPC
+from bgmi.downloader.transmission_rpc import TransmissionRPC
+from bgmi.downloader.xunlei import XunleiLixianDownload
 from bgmi.utils import print_error
 
 DOWNLOAD_DELEGATE_DICT = {
     'xunlei': XunleiLixianDownload,
     'aria2-rpc': Aria2DownloadRPC,
     'transmission-rpc': TransmissionRPC,
-    'rr!': RRDownload,
 }
